additions
current limiters, trapezoidal profile, unit corrections, and more.

+++ src/main/java/frc/robot/Gains.java
@@ -6,11 +6,11 @@ public class Gains {
 	public final double kI;
 	public final double kD;
 	public final double kFF;
-	public final int kIzone;
+	public final double kIzone;
 	public final double kMinOutput;
 	public final double kMaxOutput;
 
-	public Gains(double _kP, double _kI, double _kD, double _kFF, int _kIzone, double _kMinOutput, double _kMaxOutput) {
+	public Gains(double _kP, double _kI, double _kD, double _kFF, double _kIzone, double _kMinOutput, double _kMaxOutput) {
 
 		kP = _kP;
 		kI = _kI;
